Python 3 Support (Maya 2022 and above)
Adding support for Maya 2022 and above. Fixing the reload functions.

+++ core/ui_function.py
@@ -41,7 +41,7 @@ def maya_workspace_docker(qt_widget, window_title):
     workspace_control = pm.workspaceControl(
         window_title, tabToControl=["AttributeEditor", -1], label=label)
     workspace_pointer = omui.MQtUtil.findControl(workspace_control)
-    wrap_widget = wrapInstance(long(workspace_pointer), QtWidgets.QWidget)
+    wrap_widget = wrapInstance(int(workspace_pointer), QtWidgets.QWidget)
     wrap_widget.setAttribute(QtCore.Qt.WA_DeleteOnClose)
     # wrap_widget.destroyed.connect(close)
     child = qt_widget(wrap_widget)
